Move utils.js file to validators.js
